Fixes Windows compilation.

+++ include/reactphysics3d/collision/ConvexMesh.h
@@ -135,7 +135,7 @@ class ConvexMesh {
  * @return The number of vertices in the mesh
  */
 RP3D_FORCE_INLINE uint32 ConvexMesh::getNbVertices() const {
-    return mVertices.size();
+    return (uint32_t)mVertices.size();
 }
 
 /// Return a vertex
