+++ include/reactphysics3d/containers/Map.h
@@ -53,7 +53,7 @@ class Map {
         static constexpr float DEFAULT_LOAD_FACTOR = 0.75;
 
         /// Invalid index in the array
-        static constexpr uint64 INVALID_INDEX = -1;
+        static constexpr uint64 INVALID_INDEX = (uint64_t)-1;
 
         // -------------------- Attributes -------------------- //
 
@@ -280,7 +280,7 @@ class Map {
             assert(nbAllocatedEntries > 0);
 
             // Make sure capacity is an integral multiple of alignment
-            nbAllocatedEntries = std::ceil(nbAllocatedEntries / float(GLOBAL_ALIGNMENT)) * GLOBAL_ALIGNMENT;
+            nbAllocatedEntries = (uint64_t)std::ceil(nbAllocatedEntries / float(GLOBAL_ALIGNMENT)) * GLOBAL_ALIGNMENT;
 
             Pair<K, V>* newEntries = static_cast<Pair<K, V>*>(mAllocator.allocate(nbAllocatedEntries * sizeof(Pair<K, V>)));
             uint64* newNextEntries = static_cast<uint64*>(mAllocator.allocate(nbAllocatedEntries * sizeof(uint64)));
