+++ include/reactphysics3d/containers/Set.h
@@ -52,7 +52,7 @@ class Set {
         static constexpr float DEFAULT_LOAD_FACTOR = 0.75;
 
         /// Invalid index in the array
-        static constexpr uint64 INVALID_INDEX = -1;
+        static constexpr uint64 INVALID_INDEX = (uint64_t)-1;
 
         // -------------------- Attributes -------------------- //
 
@@ -279,7 +279,7 @@ class Set {
             assert(nbAllocatedEntries > 0);
 
             // Make sure capacity is an integral multiple of alignment
-            nbAllocatedEntries = std::ceil(nbAllocatedEntries / float(GLOBAL_ALIGNMENT)) * GLOBAL_ALIGNMENT;
+            nbAllocatedEntries = (uint64_t)std::ceil(nbAllocatedEntries / float(GLOBAL_ALIGNMENT)) * GLOBAL_ALIGNMENT;
 
             V* newEntries = static_cast<V*>(mAllocator.allocate(nbAllocatedEntries * sizeof(V)));
             uint64* newNextEntries = static_cast<uint64*>(mAllocator.allocate(nbAllocatedEntries * sizeof(uint64)));
